Move models directory
