+++ examples/react-native/v12/TestApp/src/models/Bird.ts
